+++ packages/veraswap-sdk/src/basket/index.ts
@@ -1,5 +1,6 @@
 export * from "./basket.js";
 export * from "./getBasketMint.js";
+export * from "./getBasketMintQuote.js";
 export * from "./getBasketQuotes.js";
 export * from "./getBasketSwaps.js";
 export * from "./getBatchSwaps.js";
